% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCst.R
\name{logL}
\alias{logL}
\title{calculate the log-likelihood}
\usage{
logL(theta, sk, d.times, n1.h.mat, n2.h.mat, eta.sk.mat, sk.le.dh, Ewt.die)
}
\arguments{
\item{Ewt.die}{}
}
\value{

}
\description{
calculate the log-likelihood
}
