% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCst.R
\name{main1.sub.func}
\alias{main1.sub.func}
\title{calclualte AUCst}
\usage{
main1.sub.func(
  da,
  da.long,
  sk,
  par0 = c(0.5, 0, 0, 0, 0, 0, 0, 0, 0, 0),
  tseq.eval,
  resample = 0,
  nsap = 1,
  seed = 12345
)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
calclualte AUCst
}
